+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## Version 4.4.0
 
-  - Add support for specifying the host server as VIYA_HOST environment variable. This is the url of the form <protocol>://X.Y.COM. This replaces the SAS_SSL_ENABLED and VIYA_SERVER env variables. SAS_SSL_ENABLED is superseded by the TLS enviornment variable.
+  - Support specifing VIYA_SERVER with protocol
 
   This simplifies the set up of applications.